refactor record id from type alias into a class
Seems to be working, has to test a bit more.

+++ demo/demo-lib/src/jsMain/kotlin/zakadabar/demo/frontend/lib/crud/BuiltinTable.kt
@@ -46,7 +46,7 @@ class BuiltinTable : ZkTable<BuiltinDto>() {
         + custom {
             label = t("custom")
             render = { row ->
-                if ((row.id % 2L) == 0L) {
+                if ((row.id.toLong() % 2L) == 0L) {
                     + t("odd")
                 } else {
                     + t("even")
